extracted the mouse and click behaviors to their own class
